Devis fonctionel, a voir pour gérer la suppression.

+++ Devis/Models/QuoteItem.cs
@@ -11,7 +11,7 @@ namespace Devis.Models
         public virtual int Numerotation { get; set; }
         public virtual string Reference { get; set; }
         public virtual string Label { get; set; }
-        public virtual int? Quantity { get; set; }
+        public virtual double? Quantity { get; set; }
         public string Unit { get; set; }
         public virtual double? Disbursed { get; set; }
         public virtual double? Price { get; set; }
